[Refactor] 모듈화 준비를 위한 인증/라우터 구조 개선 및 UI컴포넌트, 엔티티를 도메인 별로 분리
