+++ EatsOrder/Source/Core/Helper/UserDefault.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 @propertyWrapper
-struct UserDefault<T> {
+public struct UserDefault<T> {
   private let key: String
   private let defaultValue: T
   
@@ -17,7 +17,7 @@ struct UserDefault<T> {
     self.defaultValue = defaultValue
   }
   
-  var wrappedValue: T {
+  public var wrappedValue: T {
     get { return UserDefaults.standard.object(forKey: key) as? T ?? defaultValue }
     set { UserDefaults.standard.set(newValue, forKey: key) }
   }
