+++ EatsOrder/Source/Core/Protocol/Entity.swift
@@ -1,0 +1,10 @@
+//
+//  Entity.swift
+//  EatsOrder
+//
+//  Created by 한수빈 on 2025/06/02.
+//
+
+import Foundation
+
+public typealias Entity = Codable & Identifiable & Equatable
