+++ EatsOrder/Source/MV/Store/StoreEntity.swift
@@ -7,7 +7,81 @@
 
 import Foundation
 
-// 1. 위치 기반 주변 가게 목록 조회 응답
+struct StoreInfo: Entity {
+  let storeId: String
+  let category: String
+  let name: String
+  let close: String
+  let storeImageUrls: [String]
+  let isPicchelin: Bool
+  var isPick: Bool
+  let pickCount: Int
+  let hashTags: [String]
+  let totalRating: Double
+  let totalOrderCount: Int
+  let totalReviewCount: Int
+  let geolocation: GeoLocation
+  let distance: Double?
+  let createdAt: String
+  let updatedAt: String
+
+  var id: String { storeId }
+}
+
+struct Creator: Entity {
+  let userId: String
+  let nick: String
+  let profileImage: String?
+
+  var id: String { userId }
+}
+
+
+struct StoreDetail: Entity {
+  let storeId: String
+  let category: String?
+  let name: String?
+  let description: String?
+  let hashTags: [String]
+  let open: String?
+  let close: String?
+  let address: String?
+  let estimatedPickupTime: Int
+  let parkingGuide: String?
+  let storeImageUrls: [String]
+  let isPicchelin: Bool
+  var isPick: Bool
+  let pickCount: Int
+  let totalReviewCount: Int
+  let totalOrderCount: Int
+  let totalRating: Double
+  let creator: Creator
+  let geolocation: GeoLocation
+  let menuList: [MenuItem]
+  let createdAt: String
+  let updatedAt: String
+
+  var id: String { storeId }
+  static var empty: StoreDetail { return StoreDetail(storeId: "", category: "", name: "", description: "", hashTags: [], open: "", close: "", address: "", estimatedPickupTime: 0, parkingGuide: "", storeImageUrls: [], isPicchelin: false, isPick: false, pickCount: 0, totalReviewCount: 0, totalOrderCount: 0, totalRating: 0, creator: Creator(userId: "", nick: "", profileImage: ""), geolocation: GeoLocation(longitude: 0, latitude: 0), menuList: [], createdAt: "", updatedAt: "") }
+}
+
+struct MenuItem: Entity {
+  let menuId: String
+  let storeId: String
+  let category: String
+  let name: String
+  let description: String
+  let originInformation: String
+  let price: Int
+  let isSoldOut: Bool
+  let tags: [String]
+  let menuImageUrl: String?
+  let createdAt: String
+  let updatedAt: String
+
+  var id: String { menuId }
+}
+
 struct StoreList: Entity {
   let data: [StoreInfo]
   let nextCursor: String
@@ -15,29 +89,29 @@ struct StoreList: Entity {
   var id: String { UUID().uuidString }
 }
 
-// 3. 가게 좋아요/좋아요 취소 응답
+// 가게 좋아요/좋아요 취소
 struct StoreLike: Entity {
   let likeStatus: Bool
 
   var id: String { UUID().uuidString }
 }
 
-// 4. 가게 이름 검색 응답 (StoreListResponse와 구조가 유사하지만 next_cursor가 없음)
+// 가게 이름 검색
 struct StoreSearch: Entity {
   let data: [StoreInfo]
   var id: String { UUID().uuidString }
 }
 
-// 5. 실시간 인기 가게 조회 응답 (배열 형태로 반환)
+// 실시간 인기 가게 조회 응답 (배열 형태로 반환)
 typealias PopularStores = StoreSearch
 
-// 6. 인기 검색어 목록 조회 응답
+// 인기 검색어 목록
 struct PopularSearches: Entity {
   let data: [String]
   var id: String { UUID().uuidString }
 }
 
-// 7. 내가 좋아요한 가게 조회 응답 (StoreListResponse와 동일한 구조)
+// 내가 좋아요한 가게리스트
 typealias MyLikedStores = StoreList
 
 struct BannerInfo: Entity {
